Update slides for Physical Computing class
As I taught it in 2015-10.  I added a section on binary & bit-twiddling,
and dropped Atom in favor of nano.  I love Atom, but the remote editing
dance isn't worth it for the small amount of code in this class.

One set of slides for all 3 classes, so I can keep going if class runs
fast.  Towards the end I skip around, depending on interest.